Add files via upload
I created a code that would look for and display punctuation marks since there are so many more commas used in my text rather than periods. The red dots are the amount of periods that are used in the text and the other colored ellipses are the commas that are used in the text. I tried to figure out how to capitalize and lower case only certain parts of the text but I was only able to capitalize the entire text.